Move hbase example to hbase-addon project
