+++ stratosphere-addons/pact-hbase/src/main/java/eu/stratosphere/addons/hbase/GenericTableOutputFormat.java
@@ -13,7 +13,7 @@
  *
  **********************************************************************************************************************/
 
-package eu.stratosphere.pact.common.io;
+package eu.stratosphere.addons.hbase;
 
 import java.io.IOException;
